Add a new param type "pointer" to the ricxx API.
Traditionally the RI has supported three basic types for parameter
lists: integers, floats and strings.  Other types such as colours and
matrices are simple aggregates of these basic types.  This patch
introduces a new base type, "pointer", necessary for occasions when we
desire to pass pointers through the interface.  For example, an
application that embeds the aqsis libraries may want to supply a
callback to a display device already in memory rather than having to
implement a separate display plugin.

Details:
* Updates to the API Param / ParamList convenience functions
* Update code using Ri::TypeSpec::storageType() so that it handles the
  new pointer type correctly.
* Implement the newcore __Display_instance__ in terms of the new pointer
  type rather than using an integer as a (dangerous!) proxy.