Make DoF work with bucket rendering.
This change adapts the previous interleaved DoF sampler code to work in
the presence of bucketed rendering.  (Motion blur doesn't yet work, and
isn't connected at the API level.)  The first impression is that this
code needs quite a bit of optimization work, and it's worth checking out
the known alternative (the "interval" sampler scheme).  The main problem
appears to be that while low quality (semi-interactive) use is fairly
fast, the interleave method quickly becomes quite expensive when higher
quality settings are used.

There's a new hider parameter, "interleavewidth", used to set the width
of the interleaved sample tiles.