Bound caching for DoF rendering optimization.
A problem with using grids as the unit of allocation for bucket
rendering is that the grid bound may touch many buckets, and for any
bucket touched we need to do some work for each micropolygon in the
grid.  This is true even for micropolygons which aren't in the bucket
themselves, so it's important that we can identify such polygons and
cull them quickly.  This change speeds up such culling by caching the
geometric bound for each micropolygon inside the grid holder object.
The observed speedups are excellent (more than a factor of 2x) for DoF
rendering.

Making use of the cached bound for sampling static micropolygons is
still to be done.