Report error/warning messages with ErrorHandler
This is a step toward consistent error reporting for newcore, where we
report all error messages uniformly through the ErrorHandler interface.
As a precaution the error output is now protected by a mutex to make it
threadsafe (the message formatting happens thread locally however).

Simple log message folding is also implemented here to avoid spamming
the console with hundreds of identical error messages.