Use standard Imath:: name abbreviations
Originally the plan was to have special typedefs Vec3, Mat4, for the
float versions of the Imath::Vec3 templates.  It turns out that many of
the other Imath short vector type specializations (such as V2i) are very
useful however.  This patch removes the typedefs and instead uses the
Imath:: names since (a) they're more succinct, and (b) hopefully more
familiar to anyone who uses Imath for other things.