Fix motion blur when shutter exceeds key times.
Deforming geometry is defined by a set of key poses at given times.  If
the time interval defined by the key frames is shorter than the shutter
interval, we make the object invisible to samples outside the key frame
interval.  The other option is to use the position of the object at the
nearest key time for such samples, but this isn't as general.

Also fix a bug in determining which key interval a given sample resides.
(The bug would only have shown up for motion blur with more than one
interval.)