Motion API for bilinear patches
Implement framework for the usual MotionBegin/MotionEnd API for
geometry (only bilinear patches at this stage.)

A general method for converting primvars contained in an Ri::ParamList
into the internal PrimvarStorage class is included in
RenderApi::preparePrimvars().  This includes support for automatically
appending the color attribute to the primvar list if desired, so setting
the surface color will now work correctly.  A new function,
Geometry::motionCompatible() was added to the geometry class to allow
pieces of geometry to be tested for motion blur compatibility.