First go at point cloud based occlusion
This is very much a work in progress, but it demonstrates simple
point-based occlusion in action by baking occlusion to the point cloud
and visualizing that.  So far an array of points can be rendered to a
micro environment buffer at a given point, and the occlusion calculated
from the resulting environment depth buffer.  Caveats abound:
- No acceleration structure is implemented yet so this is _slow_
- Nearby points are inaccurately represented using squares in the
  microbuffer raster, rather than ray traced.  This gives artifacts near
  creases in a surface.
- No jittering of the microraster is used, which means banding artifacts
  in certian circumstances
- A fixed data layout [P,N,r,Cs] for the point data is assumed.