Ray tracing of nearby microdisks for point clouds
When two surfaces meet at a sharp angle, the visibility of surfels on
the adjacent surface needs to be computed more carefully than with the
usual approximate method - if not, disturbing artifacts will appear
where the surfaces join.  This patch solves the problem by resolving
visibility of nearby surfaces using ray tracing.  So far it's pretty
slow, and needs more parameter adjustment for the best quality.