Replace microdisk raytracing with rasterization
This is an optimization for the exact visibility testing part of the
microbuffer rendering.  Raytracing is simple to code, but I don't know a
good way to bound the projected ellipse on an environment map face which
results in testing more rays than necessary.  Rasterization allows the
optimal bounding box to be computed and also has a slightly lower number
of flops required per pixel tested.  The rasterization setup cost is
high, equivalent to casting several rays (probably more than 5 or so) it
will be worthwhile though, unless the raster faces are very small.