Hierarchial point cloud for microrendering
This patch speeds up microbuffer rendering using an octree as a
hierarchial representation of the point cloud.  There's plenty of
caveats: The octree construction is pretty naive, the interior nodes use
a single point to represent an aggregate of points which isn't accurate
in many situations.  The octree traversal can be controlled by an extra
command line parameter -a which sets the maximum allowed solid angle for
an interior node.

Also added the ability to centre the camera on the light probe by
pressing the 'c' key in the viewer.