Pass dicing coords explicitly to Diceable function
This allows us to specify the coordinates in which the diceable check
will occur independently from the geometry classes.  This is important
if we want to do the diceable check in a nonstandard coordinate system.
For example, we could dice in world space if we want to produce a
quality point cloud for point based lighting techniques.