Radiosity integrator for microrendering
Set up an integrator for recording radiosity in the microbuffer and
integrating it assuming a diffuse BRDF.  The PointOctree was also
extended to hold colours in internal nodes.  Internal tree nodes are now
computed using a surface area weighting factor in the averaging rather
than assuming all points are the same size.