Attribute for non raster oriented dicing.
Non raster-oriented dicing: allow objects to be diced as if all parts of the
surface face the camera.  When dicing in raster space, the object dimension
along the view direction is neglected from the calculation.  In contrast, non
raster-oriented dicing measures the dice size in a scaled version of camera
space, so the dimension along the view direction is equally important.

One motivating use case for this is in producing good quality point clouds for
point based lighting techniques, but it's also useful for avoiding displacement
artifacts on surfaces which are parallel to the viewing direction.

As for other RenderMan renderers, this feature is enabled on a per-surface
basis by setting the rasterorient attribute to zero:

	Attribute "dice" "int rasterorient" 0