Prevent infinite recursion in point octree build
When building an octree of points, if a whole bunch of the points are at
the same location it's impossible to put them into different octree
nodes.  This leads to infinite recursion unless there's some
countermeasure, and a simple (if not incredibly efficient) one is to
provide a maximum recursion depth as done here.