Clean up bake3d() & implement "interpolate" flag
In order to get good quality point clouds for point based GI, we should
interpolate the baked variables from the grid vertices to the
micropolygon centres.  This avoids duplicate rows of points at grid
boundaries, and ensures the points are fairly evenly spaced.
Interpolation is off by default, but can be enabled with the optional
"interpolate" argument to bake3d().

The bake3d() shadeop should also automatically compute the point radius,
which now works.  The optional "radius" argument also allows this to be
overridden by the user, the "radiusscale" argument lets the computed
radius be scaled.