Fix more artifacts in exact disk microrendering
The culling of nearby disks before exact micro environment rasterization
needs to be performed carefully so that disks behind the camera and
those spanning the perspective divide are handled correctly.  This
change fixes a subtle problem where disks behind the camera could
sometimes get projected in front of it for certian special sets of
parameters.  This resulted in extra spurrious occlusion and dark blotchy
artifacts.

Also revert use of doubles from previous commit, because the
rearragement gives a much more robust method to classify whether disks
span the perspective divide, so the floating point errors seem to no
longer be an issue.