Basic OpenMP parallelization of occlusion/indirectdiffuse
This is a simple minded parallelization of the inner shading loop.  It's
not a particularly good way to parallelize things, but it's very easy
and is nonintrusive to the rest of the code; it's a good temporary
measure until we get real multithreading in the new core.  Scaling with
number of threads appears to be fine for reasonable sized grids in any
case.