Bugfix #3345371: transparent grids crash motion blurred primitives
Summary: Honour request not to cull the entire grid if all of its
micropolygons are transparent.

Details:

When Aqsis renders a motion blurred object, it represents it eventually
as a CqMotionMicroPolyGrid.  This grid references a series of
CqMicroPolyGrid objects, which are the time samples.  However, only the
first of these time samples is actually shaded (in CqMicroPolyGrid::Shade).
The other time samples are kept just for their geometry.

During shading, the CqMicroPolyGrid was previously culled if all of its
micropolygons were fully transparent.  This led to problems later, when
the CqMotionMicroPolyGrid attempted to access the variables of the
culled grid.  There already existed a flag to prevent this from
happening (canCullGrid - a parameter of CqMicroPolyGrid::Shade), which
was previously ignored.

This fix appears to work, but it raises another (minor) issue: if the
CqMotionMicroPolyGrid is fully transparent (which is must be if its first
grid---the only one used for shading---is full transparent), then surely
it should be culled.  This needs further investigation.