Greatly reduce indirectdiffuse() artifacts
By rendering points in a pointcloud from front to back, the nearby
surfaces can correctly hide more distance surfaces in any pixel where
the coverage is equal to one.  Keeping track of the coverage as we go
allows us to approximately antialias edges in the microbuffer, which
greatly reduces artifacts.