Cull control attributes for point cloud generation
Implement two new attributes to allow culling to be turned off when
baking point clouds.  This is necessary since culling otherwise results
in holes in the baked point cloud.

Attribute "cull" "int backfacing" [0] allows backface culling of
micropolygon grids after displacement to be turned off so that the
surface shader still runs.

Attribute "cull" "int hidden" [0] allows culling of occluded geometry to
be turned off, to force dicing and shading of geometry which would
otherwise be discarded based only on the bounding box.  It's not clear
what to do about culling of surfaces which lie entirely outside the
viewing frustum, since these could cross the perspective divide and mess
up the calculations.  For now we just allow them to be culled, but we
probably need to do something smarter.