Port eqsl from fltk to qt4.
This is a straightforward, minimum-effort port without cleanups.  As
such various things could be done in a much more natural way using Qt
libraries rather than the combination of boost and home-grown utilities.
In particular, it would make sense to do the process management with
QProcess rather than the combination of boost::thread and CqExecute.

Also fixed a bug in the command line processing which came up during the
port.