Allow adding & removing of imagese in piqsl.
This implements adding and removal of images from the current image list
in the qt port of piqsl.  There is also some work toward drag&drop
reordering of images, but it's commented out for now since it seems like
Qt needs to serialize the image view index data for this to work
properly.