Port socket IO to qt-based piqsl
This is mostly a shuffle of code to handle socket connections out of
piqsl.cpp and into imagelistmodel.cpp - the list of images now handles
listening at the socket for new incoming image data, and the spawning of
threads for new connections.

The display doesn't yet update the image data on the screen as it comes
in.