Loading & saving of image lists in qt piqsl port
Here I've taken the old image list loading and saving code to load and
save image lists in the piqsl .bks XML format.  The obsolete fltk-based
version of the same is deleted.

Also fix a problem with updating the selection when images are loaded so
that the current selection is kept up to date, which fixes some odd
keyboard navigation problems in the image list.