Use tinyformat library for ErrorHandler formatting
This makes formatting errors with the Ri::ErrorHandler class much nicer,
and encourages error messages to be carefully formatted for clarity,
since the formatting tools are always available.  In the old code, you
had to explicitly decide when to use boost::format to make the messages
more readable which meant "never" in practise.