Fix display of odd sided images in piqsl
Due to the change to using Qt, images were sometimes displayed
incorrectly since Qt may require padding at the ends of scanlines
depending on the image size.  The display buffer is now just a QImage to
fix this problem, and converting data to the display buffer is done in
the image class.