Bugfix #3389062: aqsis -d option broken
Calling a pre-world function from the renderer RiWorld implementation
doesn't work with the new RI validation filter stuff.  This is because
the validation filter already thinks it's at world scope, so any
Ri* interface functions called by the pre-world function are rejected as
being inside the world scope.

The fix was to move the pre-world stuff into a filter which hooks onto
the WorldBegin call.  Details:

* New Ri::RendererServices::addFilter() to add filters directly rather
  than by name.
* Change Ri::Filter interface to be a bit easier to use - it can be
  linked into a filter chain after construction.  Also update the
  createFilter api etc to be ignorant of the filter chain linking
  process.
* Add new PassthroughFilter which just passes through the API calls - to
  be used as a base class.
* Remove pre-world function from aqsis.cpp, and replace with a
  PreWorldFilter class which inserts the necessary API calls into the RI
  stream just before calling WorldBegin().