Bugfix #3094582: miqser too strict to process RIB fragments
miqser doesn't process RIB fragments correctly, since the new API
validation filter was too strict with the scope checking.  The fix was
to add an "Any" scope in which all requests are valid, and make that the
starting state for miqser.

In addition this change relaxes the validation of some array lengths
in miqser for arrays which depend on the attribute state, since the
attribute state may not be known when processing RIB fragments.  (For
example, the length of bicubic RiCurve varying parameters depends on the
current basis).  Validation of array lengths is turned off completely
for RiGeometry since it's highly renderer specific.