Improvements to Ri::ParamList interface
A very typical usage of the Ri::ParamList::find*() functions seems to be
in finding types of the form "uniform type[1]", so I've adjusted the
API to make this common task as succinct as possible.  The
ParamListUsage utility class has been modified to match.

ParamListBuilder was also updated to allow construction of such uniform
types directly from an rvalue rather than requiring a pointer.