Fix recently introduced ricxx utils bugs
These bugs caused the -frames option to aqsis to no longer work:

Recent refactoring of the ParamListBuilder interface introduced some
function which interact badly with the older function signatures
[operator()(const char*, T*) is chosen over operator()(const char*, T)
when T = const char*, which isn't desired].  Simply removing the older
signatures fixes this problem.

The OnOffFilter constructor was also removed recently but this needed to
be reinstated because the member m_isActive must be properly
initialized.