Bugfix 3094580: -echoapi crashes with procedurals
This bug was due to two problems:
* RibWriter unconditionally freed procedural data after printing the
  procedural.  This left the renderer with a dangling pointer.
* RibWriter doesn't know how to print custom procedurals which have been
  dynamically loaded.

The solution taken here is to turn off all handling of procedurals for
the RibWriter instance used for echoing the API.

In order to make the output valid RIB, there's also some changes to the
TeeFilter class so that the RIB writer *End scope ending functions are
called after the main renderer's End functions.  This ensures that
WorldEnd is printed after all procedurals are expanded.