First stab at PBGI tutorial documentation.
There's a few things left to do, but this tutorial should cover
the basics of indirect lighting and ambient occlusion using point
clouds.  (Reference documentation and glossy/glow effects still need
doing.)  The location within the manual of this documentation is
probably not great - needs thought about where tutorial style sections
of the manual will be placed.

In order to present syntax highlighted RSL and RIB files, a pair of new
pygments lexers are included in renderman_lexers.py which must
unfortunately be manually installed into the pygments distribution
before attempting to build the docs with sphinx (see the top of
renderman_lexers.py for instructions).