Flush point cloud caches at end of frame
This ensures that a frame baking point cloud data may be rendered,
followed by a frame which uses that data.  The implementation is a bit
of a hack, since it relies on the caches being global.  However, it's a
simple solution for now and the same problem is present in many other
areas of the core renderer in any case.

Also improve the behaviour when a point cloud file cannot be found:
don't crash + do provide a helpful error message!