Implement texture3d with partio's kdtree lookup
This replaces the old texture3d implementation with a proper lookup
scheme based on a kdtree and filtering of nearest four points.  Since
the lookups happen without prefiltering, no effort is made to avoid
aliasing by using a filter radius based on the size of the current
shading element.

The implementation is mostly complete but still does all lookups in
"current" space whereas they should happen in world space by default and
be configurable.