PNG portability fixes + make PNG support optional
Contains a number of small fixes for portability to earlier boost
versions and for running png test cases (mktemp and unlink are not
portable).

Also fix some compiler warnings, and make the PNG support optional in
the build system - it is enabled by default, but can be turned off by
disabling the AQSIS_USE_PNG flag.